Get one last test in, to up the coverage
