+++ tests/test_sitemap_fetcher.py
@@ -1,7 +1,6 @@
 import pytest
 import requests
 import os
-from urllib.request import url2pathname
 import datetime
 from sqlalchemy.pool import StaticPool
 from sqlmodel import SQLModel, create_engine, Session
@@ -11,6 +10,7 @@ from isb_lib.sitemaps.sitemap_fetcher import (
     SitemapFileFetcher,
     ThingFetcher,
 )
+from test_utils import LocalFileAdapter
 
 
 class LocalFileSitemapIndexFetcher(SitemapIndexFetcher):
@@ -33,58 +33,6 @@ class LocalFileSitemapFileFetcher(SitemapFileFetcher):
         """Overridden to reconstruct file urls by subbing in the cwd plus the relative file path"""
         filename = os.path.join(os.getcwd(), file_url)
         return f"file://{filename}"
-
-
-# Taken from https://stackoverflow.com/questions/10123929/fetch-a-file-from-a-local-url-with-python-requests#22989322
-class LocalFileAdapter(requests.adapters.BaseAdapter):
-    """Protocol Adapter to allow Requests to GET file:// URLs"""
-
-    @staticmethod
-    def _chkpath(method, path):
-        """Return an HTTP status for the given filesystem path."""
-        if method.lower() in ("put", "delete"):
-            return 501, "Not Implemented"  # TODO
-        elif method.lower() not in ("get", "head"):
-            return 405, "Method Not Allowed"
-        elif os.path.isdir(path):
-            return 400, "Path Not A File"
-        elif not os.path.isfile(path):
-            return 404, "File Not Found"
-        elif not os.access(path, os.R_OK):
-            return 403, "Access Denied"
-        else:
-            return 200, "OK"
-
-    def send(self, req, **kwargs):  # pylint: disable=unused-argument
-        """Return the file specified by the given request
-
-        @type req: C{PreparedRequest}
-        @todo: Should I bother filling `response.headers` and processing
-               If-Modified-Since and friends using `os.stat`?
-        """
-        path = os.path.normcase(os.path.normpath(url2pathname(req.path_url)))
-        response = requests.Response()
-
-        response.status_code, response.reason = self._chkpath(req.method, path)
-        if response.status_code == 200 and req.method.lower() != "head":
-            try:
-                response.raw = open(path, "rb")
-            except (OSError, IOError) as err:
-                response.status_code = 500
-                response.reason = str(err)
-
-        if isinstance(req.url, bytes):
-            response.url = req.url.decode("utf-8")
-        else:
-            response.url = req.url
-
-        response.request = req
-        response.connection = self
-
-        return response
-
-    def close(self):
-        pass
 
 
 sitemap_fetcher_test_values = [
